# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Quiz",
        "version": "1.0"
    },
    "paths": {
        "/api/admin/answers": {
            "get": {
                "tags": [
                    "Admin - Answers"
                ],
                "summary": "Get a list of active answers",
                "description": "Returns a list of answers with status 'Y'.",
                "operationId": "getAnswerList",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Data retrieved successfully."
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "question_id": {
                                                        "type": "string",
                                                        "example": "1"
                                                    },
                                                    "answer_text": {
                                                        "type": "string",
                                                        "example": "Write a answer"
                                                    },
                                                    "is_correct": {
                                                        "type": "string",
                                                        "example": "Y"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request or no data found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 400
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Couldn't get data. Please, try again."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Database query error."
                                        },
                                        "data": {
                                            "type": "null",
                                            "example": null
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin - Answers"
                ],
                "summary": "Create a new answers",
                "description": "Create a new answers with a title",
                "operationId": "storeAnswers",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question_id",
                                    "answer_text",
                                    "is_correct"
                                ],
                                "properties": {
                                    "question_id": {
                                        "type": "string",
                                        "maxLength": 2,
                                        "example": "1"
                                    },
                                    "answer_text": {
                                        "type": "string",
                                        "maxLength": 2,
                                        "example": "Update with new answer"
                                    },
                                    "is_correct": {
                                        "type": "string",
                                        "maxLength": 1,
                                        "example": "Y/N"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Answer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "201": {
                        "description": "Answer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/answers/{id}": {
            "put": {
                "tags": [
                    "Admin - Answers"
                ],
                "summary": "Update an existing Answer",
                "description": "Update a Answer by ID with a new title",
                "operationId": "updateAnswer",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the Answers to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "question_id",
                                    "answer_text",
                                    "is_correct"
                                ],
                                "properties": {
                                    "question_id": {
                                        "type": "string",
                                        "maxLength": 2,
                                        "example": "2"
                                    },
                                    "answer_text": {
                                        "type": "string",
                                        "maxLength": 2,
                                        "example": "Write Answer"
                                    },
                                    "is_correct": {
                                        "type": "string",
                                        "maxLength": 1,
                                        "example": "Y/N"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Answer updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Answer updated successfully"
                                        },
                                        "response": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Answer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin - Answers"
                ],
                "summary": "Delete a answer",
                "description": "Deletes a answer by its ID",
                "operationId": "deleteAnswers",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the answer to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Answer deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Answer deleted successfully"
                                        },
                                        "response": {
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Answer not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in format: Bearer <token>",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}